pyproject.toml
[project]
name = "ws-agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.3",
    "faiss-cpu>=1.11.0.post1",
    "langchain-aws>=0.2.30",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.3",
    "pypdf>=5.9.0",
]

[dependency-groups]
dev = ["langgraph-cli[inmem]>=0.3.6"]
